Adds cancellation to URLSession task
